requirements.txt
Django==3.2.6
aiogram
djangorestframework
aiohttp